+++ qbraid/programs/gate_model/__init__.py
@@ -37,6 +37,7 @@ Submodules
    qasm3
    qiskit
    ionq
+   cudaq
 
 """
 import importlib