+++ qbraid/transpiler/conversions/qasm2/__init__.py
@@ -25,9 +25,10 @@ Functions
     qasm2_to_qasm3
     qasm2_to_ionq
     qasm2_to_qibo
+    qibo_to_qasm2
 
 """
-from .qasm2_extras import qasm2_to_qibo
+from .qasm2_extras import qasm2_to_qibo, qibo_to_qasm2
 from .qasm2_to_cirq import qasm2_to_cirq
 from .qasm2_to_ionq import qasm2_to_ionq
 from .qasm2_to_pytket import qasm2_to_pytket
@@ -41,4 +42,5 @@ __all__ = [
     "qasm2_to_qiskit",
     "qasm2_to_ionq",
     "qasm2_to_qibo",
+    "qibo_to_qasm2",
 ]