+++ tests/transpiler/qasm/test_qasm_to_cudaq.py
@@ -24,7 +24,7 @@ from qiskit.qasm3 import loads as qasm3_loads
 
 from qbraid.interface import assert_allclose_up_to_global_phase, circuits_allclose
 from qbraid.transpiler.conversions.openqasm3 import openqasm3_to_cudaq
-from qbraid.transpiler.conversions.openqasm3.openqasm3_extras import make_gate_kernel
+from qbraid.transpiler.conversions.openqasm3.openqasm3_to_cudaq import make_gate_kernel
 from qbraid.transpiler.conversions.qasm2.qasm2_to_qasm3 import qasm2_to_qasm3
 from qbraid.transpiler.exceptions import ProgramConversionError